Add completion commands to root
