+++ cmd/root.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"errors"
 	"os"
 
 	"github.com/civo/cli/cmd/apikey"
@@ -20,6 +21,7 @@ import (
 	"github.com/civo/cli/cmd/volume"
 	"github.com/civo/cli/common"
 	"github.com/civo/cli/config"
+	"github.com/civo/cli/utility"
 	"github.com/spf13/cobra"
 )
 
@@ -34,6 +36,76 @@ as instances and Kubernetes clusters at Civo.com.`,
 	},
 }
 
+// completionCmd represents the completion command
+var completionCmd = &cobra.Command{
+	Use:     "completion [bash|zsh|powershell|fish]",
+	Short:   "Generates bash completion scripts",
+	Example: "civo completion [bash|zsh|powershell|fish]",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		err := cmd.Help()
+		if err != nil {
+			return err
+		}
+		return errors.New("a valid subcommand is required")
+	},
+}
+
+// completionBashCmd generates bash completion scripts
+var completionBashCmd = &cobra.Command{
+	Use:   "bash",
+	Short: "Generates bash completion scripts",
+	Run: func(cmd *cobra.Command, args []string) {
+		// rootCmd.GenBashCompletion(os.Stdout)
+		err := rootCmd.GenBashCompletion(os.Stdout)
+		if err != nil {
+			utility.Error("%s", err.Error())
+			os.Exit(1)
+		}
+	},
+}
+
+// completionFishCmd generates fish completion scripts
+var completionFishCmd = &cobra.Command{
+	Use:   "fish",
+	Short: "Generates fish completion scripts",
+	Run: func(cmd *cobra.Command, args []string) {
+		// rootCmd.GenBashCompletion(os.Stdout)
+		err := rootCmd.GenFishCompletion(os.Stdout, true)
+		if err != nil {
+			utility.Error("%s", err.Error())
+			os.Exit(1)
+		}
+	},
+}
+
+// completionPowerShellCmd generates powershell completion scripts
+var completionPowerShellCmd = &cobra.Command{
+	Use:   "powershell",
+	Short: "Generates powershell completion scripts",
+	Run: func(cmd *cobra.Command, args []string) {
+		// rootCmd.GenBashCompletion(os.Stdout)
+		err := rootCmd.GenPowerShellCompletion(os.Stdout)
+		if err != nil {
+			utility.Error("%s", err.Error())
+			os.Exit(1)
+		}
+	},
+}
+
+// completionZshCmd generates zsh completion scripts
+var completionZshCmd = &cobra.Command{
+	Use:   "zsh",
+	Short: "Generates zsh completion scripts",
+	Run: func(cmd *cobra.Command, args []string) {
+		// rootCmd.GenBashCompletion(os.Stdout)
+		err := rootCmd.GenZshCompletion(os.Stdout)
+		if err != nil {
+			utility.Error("%s", err.Error())
+			os.Exit(1)
+		}
+	},
+}
+
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
@@ -45,6 +117,12 @@ func Execute() {
 
 func init() {
 	cobra.OnInitialize(config.ReadConfig)
+
+	rootCmd.AddCommand(completionCmd)
+	completionCmd.AddCommand(completionBashCmd)
+	completionCmd.AddCommand(completionZshCmd)
+	completionCmd.AddCommand(completionPowerShellCmd)
+	completionCmd.AddCommand(completionFishCmd)
 
 	rootCmd.PersistentFlags().StringVarP(&config.Filename, "config", "", "", "config file (default is $HOME/.civo.json)")
 	rootCmd.PersistentFlags().StringVarP(&common.OutputFields, "fields", "f", "", "output fields for custom format output (use -h to determine fields)")
